requirements.txt
elasticsearch
python-dotenv
google-generativeai
